# alphadash/frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "AlphaDash",
  "description": "Beginner strategy-agent product \u2014 paper-trading first.",
  "version": "0.0.1"
 },
 "paths": {
  "/auth/register": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Register",
    "operationId": "register_auth_register_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RegisterRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MeResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/auth/login": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Login",
    "operationId": "login_auth_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LoginRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MeResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/auth/me": {
   "get": {
    "tags": [
     "auth"
    ],
    "summary": "Me",
    "operationId": "me_auth_me_get",
    "parameters": [
     {
      "name": "alphadash_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Alphadash Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MeResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/orders": {
   "post": {
    "tags": [
     "orders"
    ],
    "summary": "Create Order",
    "operationId": "create_order_orders_post",
    "parameters": [
     {
      "name": "Idempotency-Key",
      "in": "header",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 8,
       "maxLength": 128,
       "title": "Idempotency-Key"
      }
     },
     {
      "name": "alphadash_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Alphadash Session"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OrderRequest"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OrderResult"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "orders"
    ],
    "summary": "List Orders",
    "operationId": "list_orders_orders_get",
    "parameters": [
     {
      "name": "alphadash_session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Alphadash Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/OrderOut"
         },
         "title": "Response List Orders Orders Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "FillOut": {
    "properties": {
     "qty": {
      "type": "string",
      "title": "Qty"
     },
     "price": {
      "type": "string",
      "title": "Price"
     },
     "fee": {
      "type": "string",
      "title": "Fee"
     },
     "filled_at": {
      "type": "string",
      "title": "Filled At"
     }
    },
    "type": "object",
    "required": [
     "qty",
     "price",
     "fee",
     "filled_at"
    ],
    "title": "FillOut"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "LoginRequest": {
    "properties": {
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     },
     "password": {
      "type": "string",
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "email",
     "password"
    ],
    "title": "LoginRequest"
   },
   "MeResponse": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "display_name": {
      "type": "string",
      "title": "Display Name"
     }
    },
    "type": "object",
    "required": [
     "id",
     "email",
     "display_name"
    ],
    "title": "MeResponse"
   },
   "OrderOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "symbol": {
      "type": "string",
      "title": "Symbol"
     },
     "asset_class": {
      "type": "string",
      "title": "Asset Class"
     },
     "side": {
      "type": "string",
      "title": "Side"
     },
     "order_type": {
      "type": "string",
      "title": "Order Type"
     },
     "qty": {
      "type": "string",
      "title": "Qty"
     },
     "limit_price": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Limit Price"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "rejected_reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rejected Reason"
     },
     "created_at": {
      "type": "string",
      "title": "Created At"
     },
     "fill": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/FillOut"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "id",
     "symbol",
     "asset_class",
     "side",
     "order_type",
     "qty",
     "limit_price",
     "status",
     "rejected_reason",
     "created_at"
    ],
    "title": "OrderOut"
   },
   "OrderRequest": {
    "properties": {
     "symbol": {
      "type": "string",
      "maxLength": 32,
      "minLength": 1,
      "title": "Symbol"
     },
     "asset_class": {
      "type": "string",
      "pattern": "^(equity|crypto)$",
      "title": "Asset Class"
     },
     "side": {
      "type": "string",
      "pattern": "^(buy|sell)$",
      "title": "Side"
     },
     "order_type": {
      "type": "string",
      "pattern": "^(market|limit)$",
      "title": "Order Type"
     },
     "qty": {
      "type": "string",
      "title": "Qty"
     },
     "limit_price": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Limit Price"
     }
    },
    "type": "object",
    "required": [
     "symbol",
     "asset_class",
     "side",
     "order_type",
     "qty"
    ],
    "title": "OrderRequest"
   },
   "OrderResult": {
    "properties": {
     "order": {
      "$ref": "#/components/schemas/OrderOut"
     },
     "violations": {
      "items": {
       "$ref": "#/components/schemas/ViolationOut"
      },
      "type": "array",
      "title": "Violations"
     },
     "replayed": {
      "type": "boolean",
      "title": "Replayed"
     }
    },
    "type": "object",
    "required": [
     "order",
     "violations",
     "replayed"
    ],
    "title": "OrderResult"
   },
...
   "RegisterRequest": {
    "properties": {
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     },
     "password": {
      "type": "string",
      "minLength": 10,
      "title": "Password"
     },
     "display_name": {
      "type": "string",
      "maxLength": 120,
      "minLength": 1,
      "title": "Display Name"
     }
    },
    "type": "object",
    "required": [
     "email",
     "password",
     "display_name"
    ],
    "title": "RegisterRequest"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "ViolationOut": {
    "properties": {
     "limit_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Limit Type"
     },
     "message": {
      "type": "string",
      "title": "Message"
     }
    },
    "type": "object",
    "required": [
     "limit_type",
     "message"
    ],
    "title": "ViolationOut"
   }
  }
 }
}